Explicitly maintain a reference to a _HandlerDelegate's task.
This ensures that, when an async task for an HTTP handler delegate
is running, a strong reference is maintained at all times by
the delegate itself.

This is a complement to https://github.com/tornadoweb/tornado/pull/3269.
While that change creates a global store of running tasks, this one
independently stores it locally.