web: Type SUPPORTED_METHODS so it can be overridden.
Its default type is `Tuple[str, str, str, str, str, str, str]`, which
can only be overridden by a tuple of the exact same length.  Which is
not terribly useful for its stated purpose.

Type it as a `Collection[str]`, which should support folks overriding
it as any number of reasonable things, as it is only used by way of
`if request.method not in self.SUPPORTED_METHODS`.