setup: Remove override of bdist_wheel
This is no longer necessary in recent versions of setuptools
(and we now check that abi3 is used properly in official releases).
The bdist_wheel implementation is being moved from the wheel package
into setuptools so overriding it will stop working in the future.

Follows the example of https://github.com/joerick/python-abi3-package-sample/pull/30

Replaces #3425